chore: extract sdk type into different files
