+++ packages/sdk/src/types/relationships/ConnectorRelationship.ts
@@ -1,23 +1,7 @@
 import { ConnectorRelationshipTemplate } from "../relationshipTemplates/ConnectorRelationshipTemplate";
 import { ConnectorIdentity } from "./ConnectorIdentity";
+import { ConnectorRelationshipAuditLogEntry } from "./ConnectorRelationshipAuditLogEntry";
 import { ConnectorRelationshipStatus } from "./ConnectorRelationshipStatus";
-
-export enum ConnectorRelationshipAuditLogEntryReason {
-    Creation = "Creation",
-    AcceptanceOfCreation = "AcceptanceOfCreation",
-    RejectionOfCreation = "RejectionOfCreation",
-    RevocationOfCreation = "RevocationOfCreation",
-    Termination = "Termination"
-}
-
-export interface ConnectorRelationshipAuditLogEntry {
-    createdAt: string;
-    createdBy: string;
-    createdByDevice: string;
-    reason: ConnectorRelationshipAuditLogEntryReason;
-    oldStatus?: ConnectorRelationshipStatus;
-    newStatus: ConnectorRelationshipStatus;
-}
 
 export interface ConnectorRelationshipAuditLog extends Array<ConnectorRelationshipAuditLogEntry> {}
