+++ packages/sdk/src/types/relationships/ConnectorRelationshipAuditLogEntryReason.ts
@@ -1,0 +1,7 @@
+export enum ConnectorRelationshipAuditLogEntryReason {
+    Creation = "Creation",
+    AcceptanceOfCreation = "AcceptanceOfCreation",
+    RejectionOfCreation = "RejectionOfCreation",
+    RevocationOfCreation = "RevocationOfCreation",
+    Termination = "Termination"
+}
